refactored questions
