+++ admin/src/main.js
@@ -18,6 +18,8 @@ import {
   faPlay,
   faArrowLeft,
   faArrowRight,
+  faArrowUp,
+  faArrowDown,
   faQuestionCircle,
   faInfoCircle,
   faChartArea,
@@ -50,6 +52,8 @@ library.add(
   faPlay,
   faArrowLeft,
   faArrowRight,
+  faArrowUp,
+  faArrowDown,
   faQuestionCircle,
   faInfoCircle,
   faChartArea,
